+++ @nerve/core/components/index.ts
@@ -1,5 +1,5 @@
 export * from './actions';
-export * from './foundations';
+export * from './cms';
 export * from './inputs';
 export * from './layout';
 export * from './media';
