+++ @nerve/domains/seo/components/StructuredData/schemas/webSiteSchema.ts
