+++ @nerve/domains/seo/components/index.ts
@@ -13,6 +13,5 @@
  */
 
 export * from './GlobalSEO/GlobalSEO';
-export * from './Link/Link';
 export * from './PageBasicSEO/PageBasicSEO';
 export * from './StructuredData/StructuredData';
