+++ @nerve/domains/seo/components/zzLegacySEO/Schema/Schema.interfaces.ts
