+++ @nerve/domains/seo/components/zzLegacySEO/Schema/__manifest__/defs/CreativeWork.js
