+++ @nerve/domains/seo/components/zzLegacySEO/Schema/__manifest__/defs/Video.js
