+++ @nerve/domains/seo/components/zzLegacySEO/Schema/__manifest__/index.js
