+++ src/pages/404.tsx
@@ -1,5 +1,5 @@
 import * as React from 'react';
-import { Container } from '@nerve/core/components/layout';
+import { Container } from '@nerve/core/components';
 import { Layout } from '@nerve/domains/app';
 
 const NoPageFound: React.FC = () => {
